exit code

+++ percycraft-smp-ami/provision.sh
@@ -20,7 +20,11 @@ provision_percycraft-smp() {
 provision_minecraft() {
     echo Provision minecraft started >&2
     mkdir -p /opt/data
-    docker compose --exit-code-from provision-minecraft -f /tmp/percycraft/percycraft-smp-ami/provision-minecraft.yml up
+    docker compose -f /tmp/percycraft/percycraft-smp-ami/provision-minecraft.yml up
+    RET=$(docker wait provision-minecraft-1 )
+    if (( $RET != 0 )); then
+        exit $RET
+    fi
     rm -rf /opt/data/.modrinth-manifest.json
     cd /opt/data
     JAR=$(ls -t *.jar | head -1)
